feat(tickets): modulo completo de gestion de entradas
- Agrega entidades Ticket, Reservation, Payment, Refund, Transfer
- Agrega enums TicketStatus, TicketType, ReservationStatus,
  PaymentStatus, RefundStatus, TransferStatus
- Implementa generacion de tickets por partido
- HU1: consultar disponibilidad de entradas
- HU2: reserva temporal con limite de 3 activas
- HU3: tiempo de retencion visible en reserva
- HU4: expiracion automatica al consultar
- HU5: confirmacion de compra con MercadoPago Sandbox
- HU6: ver todas las entradas del usuario
- Integracion con SDK MercadoPago v2.1.24

+++ backend/src/main/java/com/pjg360/PJG360/repositories/LocalFanRepository.java
@@ -3,5 +3,5 @@ package com.pjg360.PJG360.repositories;
 import com.pjg360.PJG360.model.entities.LocalFan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalFanRepository extends JpaRepository<LocalFan, Integer> {
+public interface LocalFanRepository extends JpaRepository<LocalFan, Long> {
 }
