+++ backend/src/main/java/com/pjg360/PJG360/repositories/VisitFanRepository.java
@@ -3,5 +3,5 @@ package com.pjg360.PJG360.repositories;
 import com.pjg360.PJG360.model.entities.VisitFan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VisitFanRepository  extends JpaRepository<VisitFan, Integer> {
+public interface VisitFanRepository  extends JpaRepository<VisitFan, Long> {
 }
